Add playback icon
Using two different runes from the DejaVuSans font to achieve that.
They have different "hights", so that is why there is a YOffset per rune.
Also added a small timer to know the last time we draw to the screen so we can create blinking feature (like the one introduced by this commit on the playback time)